Split `test_task.py` into two modules
